refactor: 패키지 구조 수정, 전략 도입
- 프로덕션에서는 랜덤값 기준으로 이동시키고, 테스트에서는 항상 이동시키도록 하는 MovingStrategy 인터페이스 도입
- 패키지 구조 수정
