+++ src/main/java/car_racing/domain/strategy/AlwaysMovingStrategy.java
@@ -1,0 +1,8 @@
+package car_racing.domain.strategy;
+
+public class AlwaysMovingStrategy implements CarMovingStrategy {
+    @Override
+    public boolean movable() {
+        return true;
+    }
+}
