+++ src/main/java/car_racing/domain/strategy/CarMovingStrategy.java
@@ -1,0 +1,5 @@
+package car_racing.domain.strategy;
+
+public interface CarMovingStrategy {
+    boolean movable();
+}
